Check if current story type has changed
If current story type has not changed then there is no reason to
retrieve the list of ids again. The reason is that an update on the
list could cause a bug as increments are fixed and the list could
be dynamically change by the server. Then some StoryModels could be
missed from the view, omitted by the selection range of the handler.