Fix list superposition on new request
The problem was that if a new stories request was made before the
previos one was completed, the result on the view was two merged
lists of diferent requests stories.
Implementing a global variable for story type check was the
solution and to avoid confusion the names of the related variables
were renamed.