Implement getmorecomments on child comments
Now children of top level comments have the capability to make
request to get their own children.
When a comment request is made the recursive method GetCommentTree
gets the comment object that correspond with the element that
fires the command on the view, also keeping track of the top level
parent comment in which resides.
After adding the children to the respective comment object, the
top level parent is replaced in the model with the updated element.