Add complete functionality on view
The app behaves as expected during view interaction but is evident
that changing between data contexts deletes and creates new
instances each time and that disables the posibility of mantain
data persistent on cache. The quick solution is merge the models
in one view, leaving behind the idea of changing contexts.