Enable shared cache between the contexts
Adding a singleton or static class to make only one instance of
the EndPoint class which instantiates the cache. Now the contexts
can be disposed by the window manager because the cache is
persistant.

Bug: In start, the TopStories request doesn't work but another
type of stories is requested then the TopStories work just fine
when pressing the corresponding button on the view.