Create, set and check properties on load states
Two properties were created:
- IsLoading, to indicate the state of the request.
Each time that this property is changed, raise a changed
event that set the correct style on the stories view.

- LoadingContent, to set the content of loading stories button
pointing the current state of the request.

The update on the button state is posible because of the
implementation of the RaisePropertyChanged method, that
replace the previous OnPropertyChanged method which never worked.

Also, the checks done on the button were inverted betweeen true
and false to made them more logical.

The rest of the commit are lines deleted without importance.