Add open urls capability to stories view
The view checks if an item has an url and if not, disables the
item. A click on a story title triggers the new created
GoToUrlCommand, that executes the GoToUrl method in the
StoriesViewModel which in turn sends the url to the Browser class
that checks the current OS platform and opens the url on the
default browser.