Display complete http links on comments
Before, the links displayed were the inner text of the a tags.
By default, this links in text format are cutted by the server
and because of that, the link was not completely displayed on
the view.

Now the inner text of each a tag is replaced with the href of
those tags and in consecuence can be displayed as they are on
the view.