Set user and elapsed time for each comment on view
Now, above each comment the user that post it and the elapsed
time that pass to the present moment are shown.

To support the elapsed time funcionality, it was created the
TimeConverter helper class, that gets a comment post time and
returns a string contening the amount of time that elapses from
the post moment to the present.

The string returned by the TimeConverter class stored in a new
property created in the CommentModel class that recieves a string.

On the CommentsViewModel class the result time elapsed string is
retrieved through the TimeConverter class and stored in the new
property.

In the end, the elapsed time is displayed on the view next to the
user poster.